rename classes and tests

+++ R/acs_5yr_download.R
@@ -1,6 +1,6 @@
 .onLoad <- function(libname, pkgname) {
   utils::data(
-    "dedata",
+    "acs_5yr_md",
     package = pkgname,
     envir = parent.env(environment())
   )
@@ -9,31 +9,31 @@
 # -----------------------------------------------------------------------
 
 
-#' `acs_census` S3 class
+#' `acs_5yr` S3 class
 #'
-#' A `acs_census` object is created from a given local dir. This dir will contain
+#' A `acs_5yr` object is created from a given local dir. This dir will contain
 #' the geodatabase files that we download.
 #'
 #' @param dir A string.
 #'
-#' @return A `acs_census` object.
+#' @return A `acs_5yr` object.
 #'
 #' @family data download functions
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir = dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir = dir)
 #'
 #' @export
-acs_census <- function(dir = "") {
+acs_5yr <- function(dir = "") {
   stopifnot("The dir must exist." = dir.exists(dir))
   structure(list(
     dir = dir,
-    dedata = dedata,
+    acs_5yr_md = acs_5yr_md,
     selected_files = NULL
   ),
-  class = "acs_census")
+  class = "acs_5yr")
 }
 
 
@@ -43,7 +43,7 @@ acs_census <- function(dir = "") {
 #'
 #' Gets the names of the Demographic and Economic Area Groups where data is available.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #'
 #' @return A vector, area group names.
 #'
@@ -51,8 +51,8 @@ acs_census <- function(dir = "") {
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' groups <- ac |>
 #'   get_area_groups()
@@ -63,8 +63,8 @@ get_area_groups <- function(ac)
 
 #' @rdname get_area_groups
 #' @export
-get_area_groups.acs_census<- function(ac) {
-  names(ac$dedata$groups)
+get_area_groups.acs_5yr<- function(ac) {
+  names(ac$acs_5yr_md$groups)
 }
 
 
@@ -74,7 +74,7 @@ get_area_groups.acs_census<- function(ac) {
 #'
 #' If no group is indicated, all available areas are obtained.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #' @param group A string, area group name.
 #'
 #' @return A vector, area names.
@@ -83,8 +83,8 @@ get_area_groups.acs_census<- function(ac) {
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' areas <- ac |>
 #'   get_areas(group = "Statistical Areas")
@@ -95,12 +95,12 @@ get_areas <- function(ac, group)
 
 #' @rdname get_areas
 #' @export
-get_areas.acs_census<- function(ac, group = NULL) {
-  group <- validate_names(names(ac$dedata$groups), group, 'group')
+get_areas.acs_5yr<- function(ac, group = NULL) {
+  group <- validate_names(names(ac$acs_5yr_md$groups), group, 'group')
   res <- NULL
   for (g in group) {
-    cod <- ac$dedata$groups[[g]]
-    res <- c(res, names(ac$dedata[[cod]]))
+    cod <- ac$acs_5yr_md$groups[[g]]
+    res <- c(res, names(ac$acs_5yr_md[[cod]]))
   }
   res
 }
@@ -110,7 +110,7 @@ get_areas.acs_census<- function(ac, group = NULL) {
 #'
 #' Get the years for which data has been found to be available for an area.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #' @param area A string, area name.
 #'
 #' @return A vector, area years.
@@ -119,8 +119,8 @@ get_areas.acs_census<- function(ac, group = NULL) {
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' years <- ac |>
 #'   get_area_years(area = "State")
@@ -131,18 +131,18 @@ get_area_years <- function(ac, area)
 
 #' @rdname get_area_years
 #' @export
-get_area_years.acs_census<- function(ac, area) {
+get_area_years.acs_5yr<- function(ac, area) {
   stopifnot("The area name must be defined." = !is.null(area))
   stopifnot("We can only select one area." = length(area) == 1)
-  area <- validate_names(names(ac$dedata$all_codes), area, 'area')
-  cod <- ac$dedata$all_codes[area]
-  years <- ac$dedata$years[[cod]]
+  area <- validate_names(names(ac$acs_5yr_md$all_codes), area, 'area')
+  cod <- ac$acs_5yr_md$all_codes[area]
+  years <- ac$acs_5yr_md$years[[cod]]
   last_year <- as.integer(years[length(years)])
   date <- Sys.Date()
   cur_year <- as.integer(substr(date, 1, 4))
   if (cur_year - last_year > 1) {
     for (y in (last_year + 1):(cur_year - 1)) {
-      url <- sprintf(ac$dedata$patterns$url, y, y, cod)
+      url <- sprintf(ac$acs_5yr_md$patterns$url, y, y, cod)
       if (url_file_exists(url)) {
         years <- c(years, y)
       }
@@ -157,7 +157,7 @@ get_area_years.acs_census<- function(ac, area) {
 #' Get area url file names for the given years. If no year is indicated, all
 #' available ones are obtained.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #' @param area A string, area name.
 #' @param years A vector, year number.
 #'
@@ -167,8 +167,8 @@ get_area_years.acs_census<- function(ac, area) {
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' url <- ac |>
 #'   get_area_file_names("State", 2019:2021)
@@ -182,15 +182,15 @@ get_area_file_names <- function(ac, area, years)
 
 #' @rdname get_area_file_names
 #' @export
-get_area_file_names.acs_census<- function(ac, area, years = NULL) {
+get_area_file_names.acs_5yr<- function(ac, area, years = NULL) {
   stopifnot("The area name must be defined." = !is.null(area))
   stopifnot("We can only select one area." = length(area) == 1)
-  area <- validate_names(names(ac$dedata$all_codes), area, 'area')
-  cod <- ac$dedata$all_codes[area]
-  years <- validate_names(ac$dedata$years[[cod]], years, 'year')
+  area <- validate_names(names(ac$acs_5yr_md$all_codes), area, 'area')
+  cod <- ac$acs_5yr_md$all_codes[area]
+  years <- validate_names(ac$acs_5yr_md$years[[cod]], years, 'year')
   res <- NULL
   for (y in years) {
-    res <- c(res, sprintf(ac$dedata$patterns$url, y, y, cod))
+    res <- c(res, sprintf(ac$acs_5yr_md$patterns$url, y, y, cod))
   }
   res
 }
@@ -201,18 +201,18 @@ get_area_file_names.acs_census<- function(ac, area, years = NULL) {
 #' Select area files for the given years. If no year is indicated, all available
 #' ones are selected.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #' @param area A string, area name.
 #' @param years A vector, year number.
 #'
-#' @return  A `acs_census` object.
+#' @return  A `acs_5yr` object.
 #'
 #' @family data download functions
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' ac <- ac |>
 #'   select_area_files("State", 2019:2021)
@@ -226,13 +226,13 @@ select_area_files <- function(ac, area, years)
 
 #' @rdname select_area_files
 #' @export
-select_area_files.acs_census<- function(ac, area, years = NULL) {
+select_area_files.acs_5yr<- function(ac, area, years = NULL) {
   ac$selected_files <- get_area_file_names(ac, area, years)
 
-  cod <- ac$dedata$all_codes[area]
+  cod <- ac$acs_5yr_md$all_codes[area]
   for (y in years) {
-    if (y %in% ac$dedata$too_big[[cod]]) {
-      f <- sprintf(ac$dedata$patterns$url, y, y, cod)
+    if (y %in% ac$acs_5yr_md$too_big[[cod]]) {
+      f <- sprintf(ac$acs_5yr_md$patterns$url, y, y, cod)
       cat(
         sprintf(
           "File '%s' cannot be selected, it is too heavy to be downloaded by 'utils::download.file', which is the function we are using.\n",
@@ -249,7 +249,7 @@ select_area_files.acs_census<- function(ac, area, years = NULL) {
 #'
 #' Gets the names of the files selected to be downloaded.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #'
 #' @return A vector, file names.
 #'
@@ -257,8 +257,8 @@ select_area_files.acs_census<- function(ac, area, years = NULL) {
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' groups <- ac |>
 #'   get_selected_file_names()
@@ -269,7 +269,7 @@ get_selected_file_names <- function(ac)
 
 #' @rdname get_selected_file_names
 #' @export
-get_selected_file_names.acs_census<- function(ac) {
+get_selected_file_names.acs_5yr<- function(ac) {
   ac$selected_files
 }
 
@@ -277,9 +277,9 @@ get_selected_file_names.acs_census<- function(ac) {
 #' Get too heavy file names
 #'
 #' Gets the names of the files that are too heavy to be download with the available
-#' function.
+#' function. We have downloaded them directly with the web browser.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #'
 #' @return A vector, too heavy file names.
 #'
@@ -287,8 +287,8 @@ get_selected_file_names.acs_census<- function(ac) {
 #'
 #' @examples
 #'
-#' dir <- system.file("extdata", package = "geogenr")
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' groups <- ac |>
 #'   get_too_heavy_file_names()
@@ -299,11 +299,11 @@ get_too_heavy_file_names <- function(ac)
 
 #' @rdname get_too_heavy_file_names
 #' @export
-get_too_heavy_file_names.acs_census<- function(ac) {
+get_too_heavy_file_names.acs_5yr<- function(ac) {
   res <- NULL
-  for (cod in names(ac$dedata$too_big)) {
-    for (y in ac$dedata$too_big[[cod]]) {
-      res <- c(res, sprintf(ac$dedata$patterns$url, y, y, cod))
+  for (cod in names(ac$acs_5yr_md$too_big)) {
+    for (y in ac$acs_5yr_md$too_big[[cod]]) {
+      res <- c(res, sprintf(ac$acs_5yr_md$patterns$url, y, y, cod))
     }
   }
   res
@@ -312,15 +312,15 @@ get_too_heavy_file_names.acs_census<- function(ac) {
 
 #' Download selected files
 #'
-#' Download the files that have been selected, unzip them (if desired) and, if
-#' everything went well and is indicated in the parameter, delete the downloaded
-#' files.
+#' Download the files that have been selected and have not been downloaded yet,
+#' unzip them (if desired) and, if everything went well and is indicated in the
+#' parameter, delete the downloaded files.
 #'
 #' In the `subdir` parameter, the values NULL, 'year' or 'area' can be indicated.
 #' With NULL it does not create any subdirs, with 'year' it creates them by years
 #' of downloaded files and with 'area' it creates them by areas.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #' @param subdir NULL/'year'/'area', output subdir.
 #' @param unzip A boolean, unzip files.
 #' @param delete_zip A boolean, delete zip files if correctly unzipped.
@@ -331,16 +331,14 @@ get_too_heavy_file_names.acs_census<- function(ac) {
 #'
 #' @examples
 #'
-#' dir <- tempdir()
-#' ac <- acs_census(dir)
+#' dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' ac <- acs_5yr(dir)
 #'
 #' ac <- ac |>
-#'   select_area_files("State", 2020:2021)
+#'   select_area_files("Alaska Native Regional Corporation", 2020:2021)
 #'
-#' \donttest{
 #' files <- ac |>
 #'   download_selected_files(unzip = FALSE)
-#' }
 #'
 #' @export
 download_selected_files <-
@@ -354,7 +352,9 @@ download_selected_files <-
     for (f in ac$selected_files) {
       year <- get_file_year(f)
       area <- get_file_area(f)
-      if (subdir == 'year') {
+      if (is.null(subdir)) {
+        dir <- NULL
+      } else if (subdir == 'year') {
         dir <- year
       } else if (subdir == 'area') {
         dir <- area
@@ -368,26 +368,28 @@ download_selected_files <-
       }
       exdir <- paste0(out_dir, dir)
       destfile <- paste0(exdir, basename(f))
-      res <- tryCatch(
-        utils::download.file(url = f, destfile = destfile),
-        error = function(e)
-          1
-      )
-      if (res != 1) {
-        if (unzip) {
-          res <- tryCatch(
-            utils::unzip(destfile, exdir = exdir),
-            error = function(e)
-              1
-          )
-          if (res[1] != 1) {
-            res_files <- c(res_files, gsub('.zip', '', destfile))
-            if (delete_zip) {
-              unlink(destfile)
+      if (!file.exists(destfile)) {
+        res <- tryCatch(
+          utils::download.file(url = f, destfile = destfile),
+          error = function(e)
+            1
+        )
+        if (res != 1) {
+          if (unzip) {
+            res <- tryCatch(
+              utils::unzip(destfile, exdir = exdir),
+              error = function(e)
+                1
+            )
+            if (res[1] != 1) {
+              res_files <- c(res_files, gsub('.zip', '', destfile))
+              if (delete_zip) {
+                unlink(destfile)
+              }
             }
+          } else {
+            res_files <- c(res_files, destfile)
           }
-        } else {
-          res_files <- c(res_files, destfile)
         }
       }
     }
@@ -398,13 +400,13 @@ download_selected_files <-
 #' Unzip files
 #'
 #' Unzip files that are not already unzipped in the object and, if everything
-#' went well and is indicated in the parameter, delete the zipped files.
+#' went well and is indicated in the parameter, delete the unzipped files.
 #'
 #' In the `subdir` parameter, the values NULL, 'year' or 'area' can be indicated.
 #' With NULL it does not create any subdirs, with 'year' it creates them by years
 #' of files and with 'area' it creates them by areas.
 #'
-#' @param ac A `acs_census` object.
+#' @param ac A `acs_5yr` object.
 #' @param subdir NULL/'year'/'area', output subdir.
 #' @param delete_zip A boolean, delete zip files if correctly unzipped.
 #'
@@ -415,13 +417,11 @@ download_selected_files <-
 #' @examples
 #'
 #' dir <- tempdir()
-#' ac <- acs_census(dir)
+#' source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
+#' files <- list.files(source_dir, "*.zip", full.names = TRUE)
+#' file.copy(from = files, to = dir, overwrite = TRUE)
+#' ac <- acs_5yr(dir)
 #'
-#' ac <- ac |>
-#'   select_area_files("State", 2020:2021)
-#'
-#' files <- ac |>
-#'   download_selected_files(unzip = FALSE)
 #' files <- ac |>
 #'   unzip_files()
 #'
@@ -444,7 +444,9 @@ unzip_files <- function(ac, subdir = NULL, delete_zip = FALSE) {
   res <- NULL
   for (i in 1:length(extension)) {
     f <- file[i]
-    if (subdir == 'year') {
+    if (is.null(subdir)) {
+      dir <- NULL
+    } else if (subdir == 'year') {
       dir <- get_file_year(f)
     } else if (subdir == 'area') {
       dir <- get_file_area(f)
@@ -452,16 +454,22 @@ unzip_files <- function(ac, subdir = NULL, delete_zip = FALSE) {
       dir <- NULL
     }
     if (!is.null(dir)) {
-      dir.create(file.path(mainDir = out_dir, subDir = dir), showWarnings = FALSE)
+      dir.create(file.path(mainDir = out_dir[i], subDir = dir), showWarnings = FALSE)
       dir <- paste0(dir, '/')
     }
-    exdir <- paste0(out_dir, dir)
-    dir_res <- paste0(exdir, file_name)
+    exdir <- paste0(out_dir[i], dir)
+    dir_res <- paste0(exdir, file_name[i])
     if (!dir.exists(dir_res)) {
-      utils::unzip(f, exdir = exdir)
-      res <- c(res, dir_res)
-      if (delete_zip) {
-        unlink(f)
+      r <- tryCatch(
+        utils::unzip(f, exdir = exdir),
+        error = function(e)
+          1
+      )
+      if (r[1] != 1) {
+        res <- c(res, dir_res)
+        if (delete_zip) {
+          unlink(f)
+        }
       }
     }
   }
